Add files via upload
Exercise: Pacman.

Quite happy with how it turned out. Spent a lot of time making sure the game works as intended and that the ghosts render fast enough to notice their color change when taking super food and when superfood effect is gone and the ghosts come back along with the eaten ghosts.
Didn't have time to clean up the code and minimize redundant code.
Need to add a audio sounds.
And make a proper modal for the end game.